+++ routes/class/[classId]/_layout.tsx
@@ -35,7 +35,7 @@ export default async function Layout(
             Create Post
           </a>
           <a
-            href={`/class/${classId}/faq`}
+            href={`/class/${classId}`}
             class="px-4 py-2 rounded border text-center hover:bg-gray-100"
           >
             FAQ
